RadioPropagation.jl Documentation

RadioPropagation.atmospheric_attenuation_db_per_km — Function

atmospheric_attenuation_db_per_km( frequency_ghz::Real, T_kelvin::Real=288.15, water_vapour_density_g_m³::Real=7.5, dry_air_pressure_h_pa::Real=1013.25 )::Real

Empirical model for amospheric gaseous attenuation for frequencies in the range 1 - 1000 GHz. One-way attenuation in dB/km.

Arguments

frequency_ghz The frequency of the propagating waves.

T_kelvin The absolute temperature in kelvin.

water_vapour_density_g_m³ The water vapur density in g/m³.

dry_air_pressure_h_pa The dry air pressure in hpa.

Examples

julia> res = atmospheric_attenuation_db_per_km( 22 );

julia> round(res, digits=6)
0.187337

References

Rec. ITU-R P.676-12, Attenuation by atmospheric gases and related effects, ITU-R 2020.

RadioPropagation.fog_attenuation_db_per_km — Method

fog_attenuation_db_per_km( frequency_ghz::Real, M, T_deg::Real )::Real

Empirical model for rain attenuation for frequencies above 5 GHz. One-way attenuation in dB/km.

Arguments

frequency_ghz The frequency of the propagating waves.

M The watewater concentration in g/m³.

T_deg The air temperature in degree Celsius.

Examples

julia> res = fog_attenuation_db_per_km( 10, 0.8, 23 );

julia> round(res, digits=6)
4.68976

References

M. A. Richards and J. A. Scheer and W. A. Holm, Principles of Modern Radar, SciTech Publishing, 2010.

RadioPropagation.rain_attenuation_db_per_km — Method

rain_attenuation_db_per_km( polarization::Symbol, frequency_ghz::Real, fall_rate_mm_hour::Real )::Real

Empirical model for RF rain attenuation for frequencies between 1 and 400 GHz, linear polarization. Model uses the closest frequency in the underlying data.

Arguments

polarization The polarization, :vertical, :horizontal.

frequency_ghz The frequency of the propagating waves.

fall_rate_mm_hour The rain intensity [mm/h].

Examples

julia> res = rain_attenuation_db_per_km( :vertical , 30, 20 );

julia> round(res, digits=6)
3.34

References

M. A. Richards and J. A. Scheer and W. A. Holm, Principles of Modern Radar, SciTech Publishing, 2010.

RadioPropagation.rain_attenuation_db_per_km_circular_pol — Method

rain_attenuation_db_per_km_circular_pol( frequency_ghz::Real, fall_rate_mm_hour::Real )::Real

Empirical model for rain attenuation for frequencies between 1 and 400 GHz, circular polarization. Model uses closest frequency in the underlying data. One-way attenuation in dB/km.

Arguments

frequency_ghz The frequency of the propagating waves.

fall_rate_mm_hour The rain intensity [mm/h].

Examples

julia> res = rain_attenuation_db_per_km_circular_pol( 30, 20 );

julia> round(res, digits=6)
3.675503

References

M. A. Richards and J. A. Scheer and W. A. Holm, Principles of Modern Radar, SciTech Publishing, 2010.

RadioPropagation.two_ray_propagation — Method

two_ray_propagation( distance::Real, transmit_height::Real, receive_height::Real, frequency_hz::Real, Γ::Real )::Complex

Calculate the two-way two-ray propagation between two points above a flat plane. Returns the propagation factor F. The one-way power propagation factor is |F|².

Arguments

distance The distance parallel to the plane earth.

transmit_height The height of the transmitter above the plane earth.

receive_height The height of the receiver/target above the plane earth.

Γ The reflection coefficient of the medium of the plane earth.

Examples

julia> res = two_ray_propagation( 2e3, 10, 12, 20e9, -1 );

julia> round(res, digits=6)
1.999395 + 0.034791im

References

Radar Systems Engineering Lecture 5 Propagation through the Atmosphere, IEEE New Hampshire Section IEEE AES Society, 2010